feature: add content error response for not found 404 and 400

+++ src/main/java/cl/maotech/content_service/controller/ContentController.java
@@ -3,17 +3,22 @@ package cl.maotech.content_service.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import cl.maotech.content_service.exception.ContentErrorResponse;
+import cl.maotech.content_service.exception.ContentNotFoundException;
 import cl.maotech.content_service.model.Content;
 import cl.maotech.content_service.service.ContentService;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RequestBody;
 
 @RestController
@@ -35,23 +40,51 @@ public class ContentController {
 
     @GetMapping("/{id}")
     public Content getContentById(@PathVariable Long id) {
-        return contentService.getContentById(id);
+        Content content = contentService.getContentById(id);
+        if (content == null) {
+            throw new ContentNotFoundException("Contenido no encontrado con la ID: " + id);
+        }
+        return content;
     }
 
     @PutMapping("/{id}")
     public void updateContent(@PathVariable Long id, @RequestBody Content content) {
+        Content contentToEdit = contentService.getContentById(id);
+        if (contentToEdit == null) {
+            throw new ContentNotFoundException("Contenido no encontrado con la ID: " + id);
+        }
         contentService.updateContent(id, content);
     }
 
     @DeleteMapping("/{id}")
     public void deleteContent(@PathVariable Long id) {
+        Content content = contentService.getContentById(id);
+        if (content == null) {
+            throw new ContentNotFoundException("Contenido no encontrado con la ID: " + id);
+        }
         contentService.deleteContent(id);
     }
 
     @GetMapping("/search")
     public List<Content> getContentByTypeAndStatus(
-            @RequestParam String type, 
+            @RequestParam String type,
             @RequestParam String status) {
         return contentService.getContentByTypeAndStatus(type, status);
+    }
+
+    @ExceptionHandler
+    public ResponseEntity<ContentErrorResponse> handleException(ContentNotFoundException ex) {
+        ContentErrorResponse errorResponse = new ContentErrorResponse();
+        errorResponse.setStatusCode(HttpStatus.NOT_FOUND.value());
+        errorResponse.setMessage(ex.getMessage());
+        return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
+    }
+
+    @ExceptionHandler
+    public ResponseEntity<ContentErrorResponse> handleException(Exception ex) {
+        ContentErrorResponse errorResponse = new ContentErrorResponse();
+        errorResponse.setStatusCode(HttpStatus.BAD_REQUEST.value());
+        errorResponse.setMessage(ex.getMessage());
+        return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
     }
 }
